Properly separate EmojiSearchbar components
